Fix issue that causes container to be deployed to default cluster.
bash -ex causes the launch script to bail if it encounters an error. Transient network issues caused the efs mount to fail when a DNS lookup for the efs service didn't complete. This failure happened before ECS_CLUSTER was set in ecs.config, which caused the container to be deployed to the wrong cluster, which made the cloudformation stack creation fail.

Also:
- Fix line endings
- Change EFS to a single availability zone for One Zone pricing
- Simplify GetAZs call (it gets the current region by default)
- Remove unnecessary subnetting now that we use One Zone. Be more explicit about the VPC subnet used for EFS
- Make EFS mount persistent